.history/requirements_20260310171014.txt
streamlit
sqlalchemy
pandas
openpyxl
requests
pymupdf
openai
python-dotenv
watchdog
playwright>=1.40.0
playwright-stealth>=1.0.1  # 反爬虫检测
pillow>=10.0.0